+++ Code/cyclegan_website/utils/initializeLabel.ts
@@ -1,0 +1,9 @@
+export const initializeLabel = function (
+  label: HTMLElement,
+  backgroundImage: string
+) {
+  label.style.backgroundImage = backgroundImage;
+  label.style.backgroundSize = "100% 100%";
+  label.style.backgroundRepeat = "no-repeat";
+  label.textContent = "";
+};
